Configured Project for Merge Intervals
-	Renamed Solution to ThreeSum and updated refs.
Configured Project for Merge Intervals

-	Renamed Solution to ThreeSum and updated refs.
-	Created new Solution class for merge intervals
-	Created new Solution class for merge intervals
